+++ PlanetariumRedux/Engine/Scene/Scene.cpp
@@ -53,7 +53,7 @@ Scene* Scene::getScene(int index)
 	return SceneList[index];
 }
 
-void Scene::Instantiate(EngineObject* engineObject, Vector3f position = 0, Vector3f rotation = 0)
+void Scene::Instantiate(EngineObject* engineObject, Vector3d position = 0, Vector3f rotation = 0)
 {
 	engineObject->transform.position = position;
 	engineObject->transform.Rotate(rotation);
